chore: :gear: 타입정의 파일 생성 및 이동 #6
