+++ src/types/review.ts
@@ -1,0 +1,9 @@
+export interface ReviewInfo {
+  author: string;
+  likeCount: number;
+  scrapCount: number;
+  createdAt: string;
+  content: string;
+  tags: string[];
+  id: string | number;
+}
