+++ src/types/shop.ts
@@ -1,0 +1,7 @@
+export interface ShopInfo {
+  address: string;
+  name: string;
+  category: string;
+  img: string;
+  id: string | number;
+}
